+++ packages/stream_core_flutter/lib/src/factory/stream_component_factory.g.theme.dart
@@ -64,6 +64,7 @@ mixin _$StreamComponentBuilders {
       reactionPicker: t < 0.5 ? a.reactionPicker : b.reactionPicker,
       reactions: t < 0.5 ? a.reactions : b.reactions,
       retryBadge: t < 0.5 ? a.retryBadge : b.retryBadge,
+      sheetHeader: t < 0.5 ? a.sheetHeader : b.sheetHeader,
       skeletonLoading: t < 0.5 ? a.skeletonLoading : b.skeletonLoading,
       stepper: t < 0.5 ? a.stepper : b.stepper,
       textInput: t < 0.5 ? a.textInput : b.textInput,
@@ -111,6 +112,7 @@ mixin _$StreamComponentBuilders {
     Widget Function(BuildContext, StreamReactionPickerProps)? reactionPicker,
     Widget Function(BuildContext, StreamReactionsProps)? reactions,
     Widget Function(BuildContext, StreamRetryBadgeProps)? retryBadge,
+    Widget Function(BuildContext, StreamSheetHeaderProps)? sheetHeader,
     Widget Function(BuildContext, StreamSkeletonLoadingProps)? skeletonLoading,
     Widget Function(BuildContext, StreamStepperProps)? stepper,
     Widget Function(BuildContext, StreamTextInputProps)? textInput,
@@ -155,6 +157,7 @@ mixin _$StreamComponentBuilders {
       reactionPicker: reactionPicker ?? _this.reactionPicker,
       reactions: reactions ?? _this.reactions,
       retryBadge: retryBadge ?? _this.retryBadge,
+      sheetHeader: sheetHeader ?? _this.sheetHeader,
       skeletonLoading: skeletonLoading ?? _this.skeletonLoading,
       stepper: stepper ?? _this.stepper,
       textInput: textInput ?? _this.textInput,
@@ -208,6 +211,7 @@ mixin _$StreamComponentBuilders {
       reactionPicker: other.reactionPicker,
       reactions: other.reactions,
       retryBadge: other.retryBadge,
+      sheetHeader: other.sheetHeader,
       skeletonLoading: other.skeletonLoading,
       stepper: other.stepper,
       textInput: other.textInput,
@@ -262,6 +266,7 @@ mixin _$StreamComponentBuilders {
         _other.reactionPicker == _this.reactionPicker &&
         _other.reactions == _this.reactions &&
         _other.retryBadge == _this.retryBadge &&
+        _other.sheetHeader == _this.sheetHeader &&
         _other.skeletonLoading == _this.skeletonLoading &&
         _other.stepper == _this.stepper &&
         _other.textInput == _this.textInput &&
@@ -308,6 +313,7 @@ mixin _$StreamComponentBuilders {
       _this.reactionPicker,
       _this.reactions,
       _this.retryBadge,
+      _this.sheetHeader,
       _this.skeletonLoading,
       _this.stepper,
       _this.textInput,