+++ packages/stream_core_flutter/lib/src/theme/stream_theme.dart
@@ -24,6 +24,7 @@ import 'components/stream_playback_speed_toggle_theme.dart';
 import 'components/stream_progress_bar_theme.dart';
 import 'components/stream_reaction_picker_theme.dart';
 import 'components/stream_reactions_theme.dart';
+import 'components/stream_sheet_header_theme.dart';
 import 'components/stream_skeleton_loading_theme.dart';
 import 'components/stream_stepper_theme.dart';
 import 'components/stream_switch_theme.dart';
@@ -123,6 +124,7 @@ class StreamTheme extends ThemeExtension<StreamTheme> with _$StreamTheme {
     StreamProgressBarThemeData? progressBarTheme,
     StreamReactionPickerThemeData? reactionPickerTheme,
     StreamReactionsThemeData? reactionsTheme,
+    StreamSheetHeaderThemeData? sheetHeaderTheme,
     StreamSkeletonLoadingThemeData? skeletonLoadingTheme,
     StreamStepperThemeData? stepperTheme,
     StreamSwitchThemeData? switchTheme,
@@ -163,6 +165,7 @@ class StreamTheme extends ThemeExtension<StreamTheme> with _$StreamTheme {
     progressBarTheme ??= const StreamProgressBarThemeData();
     reactionPickerTheme ??= const StreamReactionPickerThemeData();
     reactionsTheme ??= const StreamReactionsThemeData();
+    sheetHeaderTheme ??= const StreamSheetHeaderThemeData();
     skeletonLoadingTheme ??= const StreamSkeletonLoadingThemeData();
     stepperTheme ??= const StreamStepperThemeData();
     switchTheme ??= const StreamSwitchThemeData();
@@ -197,6 +200,7 @@ class StreamTheme extends ThemeExtension<StreamTheme> with _$StreamTheme {
       progressBarTheme: progressBarTheme,
       reactionPickerTheme: reactionPickerTheme,
       reactionsTheme: reactionsTheme,
+      sheetHeaderTheme: sheetHeaderTheme,
       skeletonLoadingTheme: skeletonLoadingTheme,
       stepperTheme: stepperTheme,
       switchTheme: switchTheme,
@@ -245,6 +249,7 @@ class StreamTheme extends ThemeExtension<StreamTheme> with _$StreamTheme {
     required this.progressBarTheme,
     required this.reactionPickerTheme,
     required this.reactionsTheme,
+    required this.sheetHeaderTheme,
     required this.skeletonLoadingTheme,
     required this.stepperTheme,
     required this.switchTheme,
@@ -373,6 +378,9 @@ class StreamTheme extends ThemeExtension<StreamTheme> with _$StreamTheme {
   /// The reaction theme for this theme.
   final StreamReactionsThemeData reactionsTheme;
 
+  /// The sheet header theme for this theme.
+  final StreamSheetHeaderThemeData sheetHeaderTheme;
+
   /// The skeleton theme for this theme.
   final StreamSkeletonLoadingThemeData skeletonLoadingTheme;
 
@@ -431,6 +439,7 @@ class StreamTheme extends ThemeExtension<StreamTheme> with _$StreamTheme {
       progressBarTheme: progressBarTheme,
       reactionPickerTheme: reactionPickerTheme,
       reactionsTheme: reactionsTheme,
+      sheetHeaderTheme: sheetHeaderTheme,
       skeletonLoadingTheme: skeletonLoadingTheme,
       stepperTheme: stepperTheme,
       switchTheme: switchTheme,